Minor changes made
As requested, I changed GW's to GWs and PBH's to PBHs. There was already a \pagebreak command present - there were a lot of empty lines before it, so I removed it.
About the inverted commas with variable, I haven't seen any research paper do that, so I will leave it like that.